All example PBIXs can be run through! Fixed bug where ampersands in PBIX path weren't being properly XMLA-sanitized (missing ";") Added fields/classes to layout and SSAS models as new data was processed from the example PBIXs Switched some unions to discriminated unions.

+++ pbi_core/static_files/file_classes/diagram_layout.py
@@ -59,7 +59,7 @@ class DiagramV110(BaseFileModel):
     scrollPosition: Position
     nodes: list[Node110]
     name: str
-    zoomValue: int
+    zoomValue: float
     pinKeyFieldsToTop: bool
     showExtraHeaderInfo: bool
     hideKeyFieldsWhenCollapsed: bool
